develop  doctor crud

+++ src/main/java/com/Doctor/Channeling/and/Predictive/Health/Risk/System/backend/entity/Doctor.java
@@ -87,5 +87,5 @@ public class Doctor {
 
     @Column(name = "roleId")
      @NonNull
-    private int roleId;
+    private long roleId;
 }
